group summary is functional now, and changed some floats to double since giving classcastexception

+++ opencommercesearch-api/app/org/opencommercesearch/api/models/CustomerReview.scala
@@ -33,7 +33,7 @@ import org.apache.commons.lang3.StringUtils
 
 case class CustomerReview(
   var count: Option[Int],
-  var average: Option[Float]) { 
+  var average: Option[Double]) { 
     def this() = this(None, None) 
   }
